backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
bleach==6.1.0